// app/src/pages/experiment/__generated__/ExperimentCompareTableVirtualized_comparisons.graphql.ts
export type ExperimentRunAnnotatorKind = "CODE" | "HUMAN" | "LLM";
import { FragmentRefs } from "relay-runtime";
export type ExperimentCompareTableVirtualized_comparisons$data = {
  readonly compareExperiments: {
    readonly edges: ReadonlyArray<{
      readonly comparison: {
        readonly example: {
          readonly id: string;
          readonly revision: {
            readonly input: any;
            readonly referenceOutput: any;
          };
        };
        readonly runComparisonItems: ReadonlyArray<{
          readonly experimentId: string;
          readonly runs: ReadonlyArray<{
            readonly annotations: {
              readonly edges: ReadonlyArray<{
                readonly annotation: {
                  readonly annotatorKind: ExperimentRunAnnotatorKind;
                  readonly explanation: string | null;
                  readonly id: string;
                  readonly label: string | null;
                  readonly name: string;
                  readonly score: number | null;
                  readonly trace: {
                    readonly projectId: string;
                    readonly traceId: string;
                  } | null;
                };
              }>;
            };
            readonly costSummary: {
              readonly total: {
                readonly cost: number | null;
                readonly tokens: number | null;
              };
            };
            readonly endTime: string;
            readonly error: string | null;
            readonly id: string;
            readonly output: any | null;
            readonly startTime: string;
            readonly trace: {
              readonly projectId: string;
              readonly traceId: string;
            } | null;
          }>;
        }>;
      };
    }>;
  };
  readonly dataset: {
    readonly experiments?: {
      readonly edges: ReadonlyArray<{
        readonly experiment: {
          readonly id: string;
          readonly metadata: any;
          readonly name: string;
          readonly project: {
            readonly id: string;
          } | null;
          readonly sequenceNumber: number;
        };
      }>;
    };
    readonly id: string;
  };
  readonly " $fragmentType": "ExperimentCompareTableVirtualized_comparisons";
};
export type ExperimentCompareTableVirtualized_comparisons$key = {
  readonly " $data"?: ExperimentCompareTableVirtualized_comparisons$data;
  readonly " $fragmentSpreads": FragmentRefs<"ExperimentCompareTableVirtualized_comparisons">;
};